[base] Use wrapper type for PlatformThreadId
Introduce a wrapper class (similar to base::StrongAlias) for
PlatformThreadId, whose API assumes that the value can be 64-bit. It
forces explicit conversion to integers with various bitness, and makes
conversion to 32-bit be explicit about truncation.

The values are currently all still 32-bit -- follow-up work will make
thread id values on macOS 64-bit.

Bug: 40187449
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/6206469
Owners-Override: Mark Mentovai <mark@chromium.org>
Auto-Submit: Leszek Swirski <leszeks@chromium.org>
Commit-Queue: Leszek Swirski <leszeks@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1416712}

+++ components/viz/service/debugger/viz_debugger.h
@@ -151,7 +151,10 @@ class VIZ_SERVICE_EXPORT VizDebugger {
 
   struct CallSubmitCommon {
     CallSubmitCommon() = default;
-    CallSubmitCommon(int index, int source, int thread, DrawOption draw_option)
+    CallSubmitCommon(int index,
+                     int source,
+                     int64_t thread,
+                     DrawOption draw_option)
         : draw_index(index),
           source_index(source),
           thread_id(thread),
@@ -159,7 +162,7 @@ class VIZ_SERVICE_EXPORT VizDebugger {
     base::Value::Dict GetDictionaryValue() const;
     int draw_index;
     int source_index;
-    int thread_id;
+    int64_t thread_id;
     VizDebugger::DrawOption option;
   };
 
@@ -167,7 +170,7 @@ class VIZ_SERVICE_EXPORT VizDebugger {
     DrawCall() = default;
     DrawCall(int index,
              int source,
-             int thread,
+             int64_t thread,
              DrawOption draw_option,
              gfx::SizeF size,
              gfx::Vector2dF position,
@@ -191,7 +194,7 @@ class VIZ_SERVICE_EXPORT VizDebugger {
     LogCall() = default;
     LogCall(int index,
             int source,
-            int thread,
+            int64_t thread,
             DrawOption draw_option,
             std::string str)
         : CallSubmitCommon(index, source, thread, draw_option),
